feat: Implementar sistema completo de gestão de abrigo animal
- Adicionar todos os adapters JPA para repositórios (Animal, Medication, Food, etc.)
- Implementar controllers REST para Animals e Medications
- Criar services completos com uso dos cases patterns
- Resolver conflitos de dependências e beans duplicados
- Configurar arquitetura hexagonal com separação clara de responsabilidades
- Sistema funcionando completamente com API REST testada

+++ src/main/resources/db/migration/V007__create_food_usage_table.sql
@@ -1,9 +1,8 @@
-
-CREATE TABLE food_usage (
+    CREATE TABLE food_usage (
     id BIGSERIAL PRIMARY KEY,
     food_id BIGINT NOT NULL,
     animal_id BIGINT,
-    quantity_used DECIMAL(8,2) NOT NULL,
+    quantity_used DOUBLE PRECISION NOT NULL,
     usage_date DATE NOT NULL,
     used_by VARCHAR(255),
     observations TEXT,
